+++ src/main/java/devin/task/Deadline.java
@@ -1,0 +1,33 @@
+package devin.task;
+
+import java.time.LocalDateTime;
+import java.time.format.DateTimeFormatter;
+
+public class Deadline extends Task {
+    public static DateTimeFormatter formatter1 = DateTimeFormatter.ofPattern("d/M/yyyy HHmm");
+    public static DateTimeFormatter formatter2 = DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm");
+
+    protected LocalDateTime by;
+
+    /**
+     * Constructs a new instance of Deadline with the specified description. by and isDone.
+     *
+     * @param description task description.
+     * @param by          task deadline.
+     * @param isDone      whether the task is completed or not.
+     */
+    public Deadline(String description, LocalDateTime by, boolean isDone) {
+        super(description, isDone);
+        this.by = by;
+    }
+
+    @Override
+    public String toFileString() {
+        return "D | " + super.toFileString() + " | " + by.format(formatter1);
+    }
+
+    @Override
+    public String toString() {
+        return "[D]" + super.toString() + " (by: " + by.format(formatter2) + ")";
+    }
+}
